tag-manager: Use regular initialization
Somehow we have ended up in a situation where the singleton object
is not initialized on creation but instead when requires the
NautilusApplication code to pass a GCancellable instance in order to
complete the initialization.

The passed cancellable pointer is stored by the application object
in order to cancel it on finalization, but there is no other usage
for this cancellabled from outside of the NautilusTagManager code.

So, we can just cancel it in NautilusTagManager::finalize, such that
there is no need for NautilusApplication to pass or store it.

This way, we can move the initalization from ::set_cancellable()
into ::init instead.